+++ components/AccessoriesComponent/AccessoriesComponent.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import {
   Text,
-  View
+  View,
+  TouchableOpacity
 } from 'react-native';
 import { merge } from 'lodash';
 
@@ -44,10 +45,12 @@ export class AccessoriesComponent extends Component {
     }
     
     return (
-      <View  style={style.container}>
-        <Text style={style.value}>{this.props.value}</Text>
-        <Text style={style.name}>{this.props.name}</Text>
-      </View>
+      <TouchableOpacity>
+        <View  style={style.container}>
+          <Text style={style.value}>{this.props.value}</Text>
+          <Text style={style.name}>{this.props.name}</Text>
+        </View>
+      </TouchableOpacity>
     );
   }
 }
