+++ components/store.js
@@ -15,3 +15,19 @@ const reducers = combineReducers({
 });
 const createStoreWithMiddleware = applyMiddleware(thunk)(createStore);
 export const store = createStoreWithMiddleware(reducers);
+
+import { mockRooms } from './RoomsComponent/RoomsMock';
+
+
+function fetchRooms () {
+    setTimeout(function () {
+        mockRooms.map((room, index) => {
+            store.dispatch({
+                type: 'UPDATE_PLACE',
+                place: Object.assign(room, {key: index})
+            });
+        });
+    }, 1000);
+}
+
+fetchRooms();
